Removed flexget/utils/entry.py
- Let's keep plugin specifics away from core
